% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure.R
\name{analyse_structure}
\alias{analyse_structure}
\title{Analyses hybrid structure}
\usage{
analyse_structure(name, L_sequence, R_sequence)
}
\arguments{
\item{name}{Hybrid name}

\item{L_sequence}{Left sequence}

\item{R_sequence}{Right sequence}
}
\value{
data.table with binding energy and dot-bracket structure
}
\description{
Analyses hybrid structure using \code{RNAduplex} from \code{ViennaRNA}
}
