% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_maps.R
\name{bin_matrix}
\alias{bin_matrix}
\title{Bin contact map matrix}
\usage{
bin_matrix(mat, bin.size)
}
\arguments{
\item{mat}{Contact map matrix}

\item{bin.size}{Size of bin}
}
\value{
Binnecd contact map matrix
}
\description{
Bin contact map matrix
}
