% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_maps.R
\name{bin_matrix_rect}
\alias{bin_matrix_rect}
\title{Bin a rectangular matrix}
\usage{
bin_matrix_rect(matrix_data, bin_size)
}
\arguments{
\item{matrix_data}{The rectangular matrix to bin.}

\item{bin_size}{The size of the bins}
}
\value{
A binned matrix.
}
\description{
This function bins a rectangular matrix by summing values in each bin. The bin size is adjustable.
It loops through each bin, defines the indices for the current bin, sums the values in the current bin, and stores in the binned matrix.
}
