% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hybrids.R
\name{cluster_hybrids}
\alias{cluster_hybrids}
\title{Cluster overlapping hybrids}
\usage{
cluster_hybrids(
  hybrids.dt,
  percent_overlap = 0.75,
  cluster_method = "components",
  weight = FALSE,
  fraction = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{percent_overlap}{Minimum percentage overlap for both left and right arms}

\item{hybrid.dt}{Hybrid data.table}
}
\value{
Hybrid data.table with hybrids clustered
}
\description{
Cluster overlapping hybrids using graph network analysis
}
