% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hybrids.R
\name{collapse_clusters}
\alias{collapse_clusters}
\title{Collapse clusters}
\usage{
collapse_clusters(hybrids.dt, mode = c("median", "wide"))
}
\arguments{
\item{hybrids.dt}{Hybrid data.table}
}
\value{
Clusters data.table
}
\description{
Collapse overlapping clusters of hybrids into a representative duplex
}
