% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coordinates.R
\name{convert_coordinates}
\alias{convert_coordinates}
\title{Title}
\usage{
convert_coordinates(hybrids.dt, genes.gr, filename, sam_tag = TRUE)
}
\arguments{
\item{sam_tag}{}
}
\value{

}
\description{
Title
}
