% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker.R
\name{convert_linkerbam_to_hybrids}
\alias{convert_linkerbam_to_hybrids}
\title{Title}
\usage{
convert_linkerbam_to_hybrids(
  bam,
  threads = 4,
  export_fasta = FALSE,
  fasta_filename
)
}
\arguments{
\item{bam}{BAM file}

\item{threads}{Threads}

\item{export_fasta}{Whether to export a FASTA/Q file}

\item{fasta_filename}{FASTA/Q filename}
}
\value{

}
\description{
Title
}
