% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coordinates.R
\name{export_genomic_bed}
\alias{export_genomic_bed}
\title{Export BED}
\usage{
export_genomic_bed(hybrids.dt, filename, sam_tag = TRUE)
}
\arguments{
\item{hybrids.dt}{Hybrids data.table with genomic coordinates}

\item{filename}{BED filename}

\item{sam_tag}{Add SAM tags to \code{name} column in BED12}
}
\value{
BED12 file exported
}
\description{
Calculate and export BED12 file with genomic coordinates
}
