% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractHybrids.R
\name{ExtractHybridsWithinBAM}
\alias{ExtractHybridsWithinBAM}
\title{Title}
\usage{
ExtractHybridsWithinBAM(aligned.bam, remove.repeats = FALSE)
}
\arguments{
\item{aligned.bam}{}
}
\value{

}
\description{
Title
}
