% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hybrids.R
\name{find_hybrid_overlaps_fraction}
\alias{find_hybrid_overlaps_fraction}
\title{Find overlapping pairs of hybrids}
\usage{
find_hybrid_overlaps_fraction(hybrids.dt, fraction_overlap, verbose = FALSE)
}
\arguments{
\item{hybrids.dt}{Hybrids data.table}

\item{verbose}{Print \code{bedtools} command}
}
\value{
bedpe.dt with left arm, right arm and mean percentage overlap calculated
}
\description{
Uses \code{bedtools pairtopair} to get all overlaps with fractional overlaps and then calculates percentage overlaps based on overall span of overlapping left and right arms.
Called by \code{cluster_hybrids}
}
