% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_maps.R
\name{get_contact_map}
\alias{get_contact_map}
\title{Get contact map}
\usage{
get_contact_map(hybrid.dt, genome.size, verbose = TRUE)
}
\arguments{
\item{hybrid.dt}{Hybrids data.table}

\item{genome.size}{Size of genome/gene/region}

\item{verbose}{Monitor progress}
}
\value{
Contact map matrix
}
\description{
Gets contact map matrix from hybrids.dt (previously subset to given gene/region)
}
