% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure.R
\name{get_mfe}
\alias{get_mfe}
\title{Get binding energy}
\usage{
get_mfe(name, L_sequence, R_sequence)
}
\arguments{
\item{name}{Hybrid name}

\item{L_sequence}{Left sequence}

\item{R_sequence}{Right sequence}
}
\value{
data.table with binding energy
}
\description{
Calculates binding energy using \code{RNAduplex} from \code{ViennaRNA}
}
