% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ncbi_sequence}
\alias{get_ncbi_sequence}
\title{Get sequence from NCBI}
\usage{
get_ncbi_sequence(accession)
}
\arguments{
\item{accession}{NCBI accession}
}
\value{
DNAStringSet of sequence
}
\description{
Get sequence from NCBI
}
