% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_hybrids.R
\name{load_blast8}
\alias{load_blast8}
\title{Load blast8 file}
\usage{
load_blast8(blast8)
}
\arguments{
\item{blast8}{blast8 file}
}
\value{
blast8 data.table
}
\description{
Load blast8 file as data.table
}
