% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_contact_map}
\alias{rcpp_get_contact_map}
\title{Get contact map using Rcpp}
\usage{
rcpp_get_contact_map(hybrids, gene_A_size, gene_B_size)
}
\arguments{
\item{hybrids}{The DataFrame of hybrids.}

\item{gene_A_size}{The size of gene A.}

\item{gene_B_size}{The size of gene B.}
}
\value{
An IntegerMatrix representing the contact map.
}
\description{
This function gets the contact map from a given DataFrame of hybrids. It uses Rcpp for faster computation.
It loops through each hybrid, gets the start and end points of the left and right genes, and increments the corresponding cells in the contact map.
}
