% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_functions.R
\name{reorient_hybrids}
\alias{reorient_hybrids}
\title{Reorients hybrids}
\usage{
reorient_hybrids(hybrids.dt)
}
\arguments{
\item{hybrids.dt}{Hybrids data.table}
}
\value{
Reoriented hybrids.dt
}
\description{
Reorients hybrids, so left arm always before right arm for intragenic hybrids and left arm seqname alphabetically before right arm seqname for intergenic
}
