% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure.R
\name{shuffle_sequence}
\alias{shuffle_sequence}
\title{Shuffle sequence}
\usage{
shuffle_sequence(sequence, number = 1, klet = 2, seed = 42)
}
\arguments{
\item{sequence}{Sequence as character vector}

\item{number}{Number of shuffled sequences to return}

\item{klet}{Preserve klet nucleotide frequencies}

\item{seed}{Seed}
}
\value{
Shuffled sequences as character vector
}
\description{
Shuffles sequences using \code{uShuffle}
}
