% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coordinates.R
\name{convert_coordinates}
\alias{convert_coordinates}
\title{Convert coordinates}
\usage{
convert_coordinates(hybrids.dt, genes.gr)
}
\arguments{
\item{hybrids.dt}{Hybrids data.table}

\item{genes.gr}{Transcript gtf as GRanges}
}
\value{
\code{hybrids.dt} with genomic coordinates in \code{[L|R]_genomic_*} columns
}
\description{
Convert from transcriptomic to genomic coordinates
}
