% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker.R
\name{convert_linkerbam_to_hybrids}
\alias{convert_linkerbam_to_hybrids}
\title{Title}
\usage{
convert_linkerbam_to_hybrids(
  bam,
  threads = 4,
  export_fasta = FALSE,
  fasta_filename
)
}
\arguments{
\item{fasta_filename}{}
}
\value{

}
\description{
Title
}
