% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_hybrids.R
\name{get_valid_hybrids}
\alias{get_valid_hybrids}
\title{Get valid hybrids}
\usage{
get_valid_hybrids(
  blast.query.dt,
  min_unmapped_length = 16,
  q_minoverlap = 4,
  q_maxgap = 4,
  s_minoverlap = 0,
  xlink_distance = 5,
  max_read_length = 100
)
}
\arguments{
\item{min_unmapped_length}{Minimum unmapped read length}

\item{q_minoverlap}{Maximum overlap in left and right query arms}

\item{q_maxgap}{Maximum gap between left and right query arms}

\item{s_minoverlap}{Minimum overlap in left and right subject arms}

\item{xlink_distance}{Maximum distance from start of the read (i.e. crosslink + 1 nt) before left arm starts}

\item{blast8.query.dt}{blast8 data.table with read length and metrics}
}
\value{
Hybrids data.table
}
\description{
Calculate valid hybrids from Blat partial alignments
}
