% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_hybrids.R
\name{get_valid_hybrids}
\alias{get_valid_hybrids}
\title{Title}
\usage{
get_valid_hybrids(
  blast.query.dt,
  min_unmapped_length = 16,
  q_minoverlap = 4,
  q_maxgap = 4,
  s_minoverlap = 0,
  xlink_distance = 5,
  max_read_length = 100
)
}
\arguments{
\item{xlink_distance}{}
}
\value{

}
\description{
Title
}
